Monarch API/requirements.txt
fastapi
uvicorn
monarchmoneycommunity
certifi
